@@ -54,7 +54,7 @@ class AnimateCameraState extends State<AnimateCamera> {
           children: <Widget>[
             Column(
               children: <Widget>[
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.newCameraPosition(
@@ -69,7 +69,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('newCameraPosition'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.newLatLng(
@@ -79,7 +79,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('newLatLng'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.newLatLngBounds(
@@ -93,7 +93,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('newLatLngBounds'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.newLatLngZoom(
@@ -104,7 +104,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('newLatLngZoom'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.scrollBy(150.0, -225.0),
@@ -116,7 +116,7 @@ class AnimateCameraState extends State<AnimateCamera> {
             ),
             Column(
               children: <Widget>[
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.zoomBy(
@@ -127,7 +127,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('zoomBy with focus'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.zoomBy(-0.5),
@@ -135,7 +135,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('zoomBy'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.zoomIn(),
@@ -143,7 +143,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('zoomIn'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.zoomOut(),
@@ -151,7 +151,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   },
                   child: const Text('zoomOut'),
                 ),
-                FlatButton(
+                TextButton(
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.zoomTo(16.0),
