@@ -99,17 +99,17 @@ class _MyAppState extends State<MyApp> {
               mainAxisAlignment: MainAxisAlignment.spaceAround,
               children: <Widget>[
                 Text('Can check biometrics: $_canCheckBiometrics\n'),
-                RaisedButton(
+                ElevatedButton(
                   child: const Text('Check biometrics'),
                   onPressed: _checkBiometrics,
                 ),
                 Text('Available biometrics: $_availableBiometrics\n'),
-                RaisedButton(
+                ElevatedButton(
                   child: const Text('Get available biometrics'),
                   onPressed: _getAvailableBiometrics,
                 ),
                 Text('Current State: $_authorized\n'),
-                RaisedButton(
+                ElevatedButton(
                   child: Text(_isAuthenticating ? 'Cancel' : 'Authenticate'),
                   onPressed:
                       _isAuthenticating ? _cancelAuthentication : _authenticate,
