@@ -10,6 +10,7 @@ import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugins.camera.CameraPermissions.PermissionsRegistry;
+import io.flutter.plugins.camera.types.ExposureMode;
 import io.flutter.plugins.camera.types.FlashMode;
 import io.flutter.view.TextureRegistry;
 import java.util.HashMap;
@@ -138,6 +139,73 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
           }
           break;
         }
+      case "setExposureMode":
+        {
+          String modeStr = call.argument("mode");
+          ExposureMode mode = ExposureMode.getValueForString(modeStr);
+          if (mode == null) {
+            result.error("setExposureModeFailed", "Unknown exposure mode " + modeStr, null);
+            return;
+          }
+          try {
+            camera.setExposureMode(result, mode);
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "setExposurePoint":
+        {
+          Boolean reset = call.argument("reset");
+          Double x = null;
+          Double y = null;
+          if (reset == null || !reset) {
+            x = call.argument("x");
+            y = call.argument("y");
+          }
+          try {
+            camera.setExposurePoint(result, x, y);
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "getMinExposureOffset":
+        {
+          try {
+            result.success(camera.getMinExposureOffset());
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "getMaxExposureOffset":
+        {
+          try {
+            result.success(camera.getMaxExposureOffset());
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "getExposureOffsetStepSize":
+        {
+          try {
+            result.success(camera.getExposureOffsetStepSize());
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "setExposureOffset":
+        {
+          try {
+            camera.setExposureOffset(result, call.argument("offset"));
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
       case "startImageStream":
         {
           try {
@@ -151,7 +219,7 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
       case "stopImageStream":
         {
           try {
-            camera.startPreview();
+            camera.stopImageStream();
             result.success(null);
           } catch (Exception e) {
             handleException(e, result);
