@@ -2,16 +2,26 @@ package io.flutter.plugins.camera.types;
 
 // Mirrors flash_mode.dart
 public enum FlashMode {
-  off,
-  auto,
-  always,
-  torch;
+  off("off"),
+  auto("auto"),
+  always("always"),
+  torch("torch");
+
+  private final String strValue;
+
+  FlashMode(String strValue) {
+    this.strValue = strValue;
+  }
 
   public static FlashMode getValueForString(String modeStr) {
-    try {
-      return valueOf(modeStr);
-    } catch (IllegalArgumentException | NullPointerException e) {
-      return null;
+    for (FlashMode value : values()) {
+      if (value.strValue.equals(modeStr)) return value;
     }
+    return null;
+  }
+
+  @Override
+  public String toString() {
+    return strValue;
   }
 }
