@@ -2,7 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-/// Generic Event coming from the native side of Camera.
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
+
+import '../../camera_platform_interface.dart';
+
+/// Generic Event coming from the native side of Camera,
+/// related to a specific camera module.
 ///
 /// All [CameraEvent]s contain the `cameraId` that originated the event. This
 /// should never be `null`.
@@ -45,6 +50,18 @@ class CameraInitializedEvent extends CameraEvent {
   /// The height of the preview in pixels.
   final double previewHeight;
 
+  /// The default exposure mode
+  final ExposureMode exposureMode;
+
+  /// The default focus mode
+  final FocusMode focusMode;
+
+  /// Whether setting exposure points is supported.
+  final bool exposurePointSupported;
+
+  /// Whether setting focus points is supported.
+  final bool focusPointSupported;
+
   /// Build a CameraInitialized event triggered from the camera represented by
   /// `cameraId`.
   ///
@@ -53,14 +70,22 @@ class CameraInitializedEvent extends CameraEvent {
   CameraInitializedEvent(
     int cameraId,
     this.previewWidth,
-    this.previewHeight,
-  ) : super(cameraId);
+    this.previewHeight, [
+    this.exposureMode,
+    this.exposurePointSupported = false,
+    this.focusMode,
+    this.focusPointSupported = false,
+  ]) : super(cameraId);
 
   /// Converts the supplied [Map] to an instance of the [CameraInitializedEvent]
   /// class.
   CameraInitializedEvent.fromJson(Map<String, dynamic> json)
       : previewWidth = json['previewWidth'],
         previewHeight = json['previewHeight'],
+        exposureMode = deserializeExposureMode(json['exposureMode']),
+        exposurePointSupported = json['exposurePointSupported'] ?? false,
+        focusMode = deserializeFocusMode(json['focusMode']),
+        focusPointSupported = json['focusPointSupported'] ?? false,
         super(json['cameraId']);
 
   /// Converts the [CameraInitializedEvent] instance into a [Map] instance that
@@ -69,6 +94,10 @@ class CameraInitializedEvent extends CameraEvent {
         'cameraId': cameraId,
         'previewWidth': previewWidth,
         'previewHeight': previewHeight,
+        'exposureMode': serializeExposureMode(exposureMode),
+        'exposurePointSupported': exposurePointSupported,
+        'focusMode': serializeFocusMode(focusMode),
+        'focusPointSupported': focusPointSupported,
       };
 
   @override
@@ -78,11 +107,21 @@ class CameraInitializedEvent extends CameraEvent {
           other is CameraInitializedEvent &&
           runtimeType == other.runtimeType &&
           previewWidth == other.previewWidth &&
-          previewHeight == other.previewHeight;
+          previewHeight == other.previewHeight &&
+          exposureMode == other.exposureMode &&
+          exposurePointSupported == other.exposurePointSupported &&
+          focusMode == other.focusMode &&
+          focusPointSupported == other.focusPointSupported;
 
   @override
   int get hashCode =>
-      super.hashCode ^ previewWidth.hashCode ^ previewHeight.hashCode;
+      super.hashCode ^
+      previewWidth.hashCode ^
+      previewHeight.hashCode ^
+      exposureMode.hashCode ^
+      exposurePointSupported.hashCode ^
+      focusMode.hashCode ^
+      focusPointSupported.hashCode;
 }
 
 /// An event fired when the resolution preset of the camera has changed.
