@@ -3,6 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:camera_platform_interface/src/types/exposure_mode.dart';
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -10,11 +12,16 @@ void main() {
 
   group('CameraInitializedEvent tests', () {
     test('Constructor should initialize all properties', () {
-      final event = CameraInitializedEvent(1, 1024, 640);
+      final event = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
 
       expect(event.cameraId, 1);
       expect(event.previewWidth, 1024);
       expect(event.previewHeight, 640);
+      expect(event.exposureMode, ExposureMode.auto);
+      expect(event.focusMode, FocusMode.auto);
+      expect(event.exposurePointSupported, true);
+      expect(event.focusPointSupported, true);
     });
 
     test('fromJson should initialize all properties', () {
@@ -22,57 +29,120 @@ void main() {
         'cameraId': 1,
         'previewWidth': 1024.0,
         'previewHeight': 640.0,
+        'exposureMode': 'auto',
+        'exposurePointSupported': true,
+        'focusMode': 'auto',
+        'focusPointSupported': true
       });
 
       expect(event.cameraId, 1);
       expect(event.previewWidth, 1024);
       expect(event.previewHeight, 640);
+      expect(event.exposureMode, ExposureMode.auto);
+      expect(event.exposurePointSupported, true);
+      expect(event.focusMode, FocusMode.auto);
+      expect(event.focusPointSupported, true);
     });
 
     test('toJson should return a map with all fields', () {
-      final event = CameraInitializedEvent(1, 1024, 640);
+      final event = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
 
       final jsonMap = event.toJson();
 
-      expect(jsonMap.length, 3);
+      expect(jsonMap.length, 7);
       expect(jsonMap['cameraId'], 1);
       expect(jsonMap['previewWidth'], 1024);
       expect(jsonMap['previewHeight'], 640);
+      expect(jsonMap['exposureMode'], 'auto');
+      expect(jsonMap['exposurePointSupported'], true);
+      expect(jsonMap['focusMode'], 'auto');
+      expect(jsonMap['focusPointSupported'], true);
     });
 
     test('equals should return true if objects are the same', () {
-      final firstEvent = CameraInitializedEvent(1, 1024, 640);
-      final secondEvent = CameraInitializedEvent(1, 1024, 640);
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
 
       expect(firstEvent == secondEvent, true);
     });
 
     test('equals should return false if cameraId is different', () {
-      final firstEvent = CameraInitializedEvent(1, 1024, 640);
-      final secondEvent = CameraInitializedEvent(2, 1024, 640);
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          2, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
 
       expect(firstEvent == secondEvent, false);
     });
 
     test('equals should return false if previewWidth is different', () {
-      final firstEvent = CameraInitializedEvent(1, 1024, 640);
-      final secondEvent = CameraInitializedEvent(1, 2048, 640);
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 2048, 640, ExposureMode.auto, true, FocusMode.auto, true);
 
       expect(firstEvent == secondEvent, false);
     });
 
     test('equals should return false if previewHeight is different', () {
-      final firstEvent = CameraInitializedEvent(1, 1024, 640);
-      final secondEvent = CameraInitializedEvent(1, 1024, 980);
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 980, ExposureMode.auto, true, FocusMode.auto, true);
+
+      expect(firstEvent == secondEvent, false);
+    });
+
+    test('equals should return false if exposureMode is different', () {
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.locked, true, FocusMode.auto, true);
+
+      expect(firstEvent == secondEvent, false);
+    });
+
+    test('equals should return false if exposurePointSupported is different',
+        () {
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, false, FocusMode.auto, true);
+
+      expect(firstEvent == secondEvent, false);
+    });
+
+    test('equals should return false if focusMode is different', () {
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.locked, true);
+
+      expect(firstEvent == secondEvent, false);
+    });
+
+    test('equals should return false if focusPointSupported is different', () {
+      final firstEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
+      final secondEvent = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, false);
 
       expect(firstEvent == secondEvent, false);
     });
 
     test('hashCode should match hashCode of all properties', () {
-      final event = CameraInitializedEvent(1, 1024, 640);
+      final event = CameraInitializedEvent(
+          1, 1024, 640, ExposureMode.auto, true, FocusMode.auto, true);
       final expectedHashCode = event.cameraId.hashCode ^
           event.previewWidth.hashCode ^
-          event.previewHeight.hashCode;
+          event.previewHeight.hashCode ^
+          event.exposureMode.hashCode ^
+          event.exposurePointSupported.hashCode ^
+          event.focusMode.hashCode ^
+          event.focusPointSupported.hashCode;
 
       expect(event.hashCode, expectedHashCode);
     });
