@@ -3,11 +3,15 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:math';
 
 import 'package:async/async.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:camera_platform_interface/src/events/device_event.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
 import 'package:camera_platform_interface/src/utils/utils.dart';
+import 'package:flutter/services.dart' hide DeviceOrientation;
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -24,7 +28,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1}
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              }
             });
         final camera = MethodChannelCamera();
 
@@ -107,7 +114,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1},
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              },
               'initialize': null
             });
         final camera = MethodChannelCamera();
@@ -118,8 +128,15 @@ void main() {
 
         // Act
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
-        camera.cameraEventStreamController
-            .add(CameraInitializedEvent(cameraId, 1920, 1080));
+        camera.cameraEventStreamController.add(CameraInitializedEvent(
+          cameraId,
+          1920,
+          1080,
+          ExposureMode.auto,
+          true,
+          FocusMode.auto,
+          true,
+        ));
         await initializeFuture;
 
         // Assert
@@ -130,6 +147,7 @@ void main() {
             'initialize',
             arguments: {
               'cameraId': 1,
+              'imageFormatGroup': 'unknown',
             },
           ),
         ]);
@@ -151,8 +169,15 @@ void main() {
           ResolutionPreset.high,
         );
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
-        camera.cameraEventStreamController
-            .add(CameraInitializedEvent(cameraId, 1920, 1080));
+        camera.cameraEventStreamController.add(CameraInitializedEvent(
+          cameraId,
+          1920,
+          1080,
+          ExposureMode.auto,
+          true,
+          FocusMode.auto,
+          true,
+        ));
         await initializeFuture;
 
         // Act
@@ -188,8 +213,15 @@ void main() {
           ResolutionPreset.high,
         );
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
-        camera.cameraEventStreamController
-            .add(CameraInitializedEvent(cameraId, 1920, 1080));
+        camera.cameraEventStreamController.add(CameraInitializedEvent(
+          cameraId,
+          1920,
+          1080,
+          ExposureMode.auto,
+          true,
+          FocusMode.auto,
+          true,
+        ));
         await initializeFuture;
       });
 
@@ -200,8 +232,16 @@ void main() {
         final streamQueue = StreamQueue(eventStream);
 
         // Emit test events
-        final event = CameraInitializedEvent(cameraId, 3840, 2160);
-        await camera.handleMethodCall(
+        final event = CameraInitializedEvent(
+          cameraId,
+          3840,
+          2160,
+          ExposureMode.auto,
+          true,
+          FocusMode.auto,
+          true,
+        );
+        await camera.handleCameraMethodCall(
             MethodCall('initialized', event.toJson()), cameraId);
 
         // Assert
@@ -220,13 +260,13 @@ void main() {
         // Emit test events
         final fhdEvent = CameraResolutionChangedEvent(cameraId, 1920, 1080);
         final uhdEvent = CameraResolutionChangedEvent(cameraId, 3840, 2160);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', fhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', uhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', fhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', uhdEvent.toJson()), cameraId);
 
         // Assert
@@ -247,11 +287,11 @@ void main() {
 
         // Emit test events
         final event = CameraClosingEvent(cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
 
         // Assert
@@ -270,12 +310,36 @@ void main() {
 
         // Emit test events
         final event = CameraErrorEvent(cameraId, 'Error Description');
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
+
+        // Assert
+        expect(await streamQueue.next, event);
+        expect(await streamQueue.next, event);
+        expect(await streamQueue.next, event);
+
+        // Clean up
+        await streamQueue.cancel();
+      });
+
+      test('Should receive device orientation change events', () async {
+        // Act
+        final eventStream = camera.onDeviceOrientationChanged();
+        final streamQueue = StreamQueue(eventStream);
+
+        // Emit test events
+        final event =
+            DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
 
         // Assert
         expect(await streamQueue.next, event);
@@ -304,8 +368,17 @@ void main() {
           ResolutionPreset.high,
         );
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
-        camera.cameraEventStreamController
-            .add(CameraInitializedEvent(cameraId, 1920, 1080));
+        camera.cameraEventStreamController.add(
+          CameraInitializedEvent(
+            cameraId,
+            1920,
+            1080,
+            ExposureMode.auto,
+            true,
+            FocusMode.auto,
+            true,
+          ),
+        );
         await initializeFuture;
       });
 
@@ -411,7 +484,29 @@ void main() {
         expect(channel.log, <Matcher>[
           isMethodCall('startVideoRecording', arguments: {
             'cameraId': cameraId,
+            'maxVideoDuration': null,
           }),
+        ]);
+      });
+
+      test('Should pass maxVideoDuration when starting recording a video',
+          () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'startVideoRecording': null},
+        );
+
+        // Act
+        await camera.startVideoRecording(
+          cameraId,
+          maxVideoDuration: Duration(seconds: 10),
+        );
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('startVideoRecording',
+              arguments: {'cameraId': cameraId, 'maxVideoDuration': 10000}),
         ]);
       });
 
@@ -496,6 +591,179 @@ void main() {
         ]);
       });
 
+      test('Should set the exposure mode', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setExposureMode': null},
+        );
+
+        // Act
+        await camera.setExposureMode(cameraId, ExposureMode.auto);
+        await camera.setExposureMode(cameraId, ExposureMode.locked);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setExposureMode',
+              arguments: {'cameraId': cameraId, 'mode': 'auto'}),
+          isMethodCall('setExposureMode',
+              arguments: {'cameraId': cameraId, 'mode': 'locked'}),
+        ]);
+      });
+
+      test('Should set the exposure point', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setExposurePoint': null},
+        );
+
+        // Act
+        await camera.setExposurePoint(cameraId, Point<double>(0.5, 0.5));
+        await camera.setExposurePoint(cameraId, null);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setExposurePoint', arguments: {
+            'cameraId': cameraId,
+            'x': 0.5,
+            'y': 0.5,
+            'reset': false
+          }),
+          isMethodCall('setExposurePoint', arguments: {
+            'cameraId': cameraId,
+            'x': null,
+            'y': null,
+            'reset': true
+          }),
+        ]);
+      });
+
+      test('Should get the min exposure offset', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'getMinExposureOffset': 2.0},
+        );
+
+        // Act
+        final minExposureOffset = await camera.getMinExposureOffset(cameraId);
+
+        // Assert
+        expect(minExposureOffset, 2.0);
+        expect(channel.log, <Matcher>[
+          isMethodCall('getMinExposureOffset', arguments: {
+            'cameraId': cameraId,
+          }),
+        ]);
+      });
+
+      test('Should get the max exposure offset', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'getMaxExposureOffset': 2.0},
+        );
+
+        // Act
+        final maxExposureOffset = await camera.getMaxExposureOffset(cameraId);
+
+        // Assert
+        expect(maxExposureOffset, 2.0);
+        expect(channel.log, <Matcher>[
+          isMethodCall('getMaxExposureOffset', arguments: {
+            'cameraId': cameraId,
+          }),
+        ]);
+      });
+
+      test('Should get the exposure offset step size', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'getExposureOffsetStepSize': 0.25},
+        );
+
+        // Act
+        final stepSize = await camera.getExposureOffsetStepSize(cameraId);
+
+        // Assert
+        expect(stepSize, 0.25);
+        expect(channel.log, <Matcher>[
+          isMethodCall('getExposureOffsetStepSize', arguments: {
+            'cameraId': cameraId,
+          }),
+        ]);
+      });
+
+      test('Should set the exposure offset', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setExposureOffset': 0.6},
+        );
+
+        // Act
+        final actualOffset = await camera.setExposureOffset(cameraId, 0.5);
+
+        // Assert
+        expect(actualOffset, 0.6);
+        expect(channel.log, <Matcher>[
+          isMethodCall('setExposureOffset', arguments: {
+            'cameraId': cameraId,
+            'offset': 0.5,
+          }),
+        ]);
+      });
+
+      test('Should set the focus mode', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusMode': null},
+        );
+
+        // Act
+        await camera.setFocusMode(cameraId, FocusMode.auto);
+        await camera.setFocusMode(cameraId, FocusMode.locked);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'auto'}),
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'locked'}),
+        ]);
+      });
+
+      test('Should set the exposure point', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusPoint': null},
+        );
+
+        // Act
+        await camera.setFocusPoint(cameraId, Point<double>(0.5, 0.5));
+        await camera.setFocusPoint(cameraId, null);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': 0.5,
+            'y': 0.5,
+            'reset': false
+          }),
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': null,
+            'y': null,
+            'reset': true
+          }),
+        ]);
+      });
+
       test('Should build a texture widget as preview widget', () async {
         // Act
         Widget widget = camera.buildPreview(cameraId);
@@ -509,7 +777,9 @@ void main() {
           () {
         final camera = MethodChannelCamera();
 
-        expect(() => camera.handleMethodCall(MethodCall('unknown_method'), 1),
+        expect(
+            () =>
+                camera.handleCameraMethodCall(MethodCall('unknown_method'), 1),
             throwsA(isA<MissingPluginException>()));
       });
 
@@ -589,6 +859,41 @@ void main() {
                 .having((e) => e.code, 'code', 'ZOOM_ERROR')
                 .having((e) => e.description, 'description',
                     'Illegal zoom error')));
+      });
+
+      test('Should lock the capture orientation', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'lockCaptureOrientation': null},
+        );
+
+        // Act
+        await camera.lockCaptureOrientation(
+            cameraId, DeviceOrientation.portraitUp);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('lockCaptureOrientation',
+              arguments: {'cameraId': cameraId, 'orientation': 'portraitUp'}),
+        ]);
+      });
+
+      test('Should unlock the capture orientation', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'unlockCaptureOrientation': null},
+        );
+
+        // Act
+        await camera.unlockCaptureOrientation(cameraId);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('unlockCaptureOrientation',
+              arguments: {'cameraId': cameraId}),
+        ]);
       });
     });
   });
